+++ service/src/test/java/bio/terra/pipelines/stairway/steps/imputation/PrepareImputationInputsStepTest.java
@@ -10,7 +10,7 @@ import bio.terra.pipelines.common.utils.PipelinesEnum;
 import bio.terra.pipelines.db.entities.PipelineInputDefinition;
 import bio.terra.pipelines.db.repositories.PipelineRunsRepository;
 import bio.terra.pipelines.db.repositories.PipelinesRepository;
-import bio.terra.pipelines.service.PipelinesService;
+import bio.terra.pipelines.service.PipelineInputsOutputsService;
 import bio.terra.pipelines.stairway.flights.imputation.ImputationJobMapKeys;
 import bio.terra.pipelines.testutils.BaseEmbeddedDbTest;
 import bio.terra.pipelines.testutils.StairwayTestUtils;
@@ -33,7 +33,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 class PrepareImputationInputsStepTest extends BaseEmbeddedDbTest {
 
-  @Mock PipelinesService pipelinesService;
+  @Mock PipelineInputsOutputsService pipelineInputsOutputsService;
   @Autowired PipelinesRepository pipelinesRepository;
   @Autowired ImputationConfiguration imputationConfiguration;
   @Autowired PipelineRunsRepository pipelineRunsRepository;
@@ -94,7 +94,7 @@ class PrepareImputationInputsStepTest extends BaseEmbeddedDbTest {
 
     // mock the service call to format the pipeline inputs
     Map<String, Object> fakeFormattedPipelineInputs = new HashMap<>(Map.of("foo", "bar"));
-    when(pipelinesService.formatPipelineInputs(
+    when(pipelineInputsOutputsService.constructPipelineInputs(
             testJobId,
             testPipelineInputsDefinitionList,
             TestUtils.TEST_PIPELINE_INPUTS_ARRAY_IMPUTATION,
@@ -108,7 +108,7 @@ class PrepareImputationInputsStepTest extends BaseEmbeddedDbTest {
 
     // do the step
     var prepareImputationInputsStep =
-        new PrepareImputationInputsStep(pipelinesService, imputationConfiguration);
+        new PrepareImputationInputsStep(pipelineInputsOutputsService, imputationConfiguration);
     var result = prepareImputationInputsStep.doStep(flightContext);
 
     assertEquals(StepStatus.STEP_RESULT_SUCCESS, result.getStepStatus());
@@ -126,7 +126,7 @@ class PrepareImputationInputsStepTest extends BaseEmbeddedDbTest {
   @Test
   void undoStepSuccess() {
     var prepareImputationInputsStep =
-        new PrepareImputationInputsStep(pipelinesService, imputationConfiguration);
+        new PrepareImputationInputsStep(pipelineInputsOutputsService, imputationConfiguration);
     var result = prepareImputationInputsStep.undoStep(flightContext);
 
     assertEquals(StepStatus.STEP_RESULT_SUCCESS, result.getStepStatus());
